Only updates WSLU if it needed

+++ linux_files/upgrade.sh
@@ -20,12 +20,15 @@ sudo rm -f /var/lib/rpm/.rpm.lock
 sudo dnf -y update
 sudo rm -f /var/lib/rpm/.rpm.lock
 
-(
-  source /etc/os-release && sudo dnf -y copr enable wslutilities/wslu "${ID_LIKE}"-"${VERSION_ID}"-"$(uname -m)"
-)
-sudo rm -f /var/lib/rpm/.rpm.lock
-sudo dnf -y update wslu
-sudo rm -f /var/lib/rpm/.rpm.lock
+# WSLU 3 is not installed
+if [ "$(wslsys -v | grep -c "3\.")" -eq 0 ]; then
+  (
+    source /etc/os-release && sudo dnf -y copr enable wslutilities/wslu "${ID_LIKE}"-"${VERSION_ID}"-"$(uname -m)"
+  )
+  sudo rm -f /var/lib/rpm/.rpm.lock
+  sudo dnf -y update wslu
+  sudo rm -f /var/lib/rpm/.rpm.lock
+fi
 
 # Update the release and main startup script files
 sudo curl -f https://raw.githubusercontent.com/WhitewaterFoundry/Fedora-Remix-for-WSL/master/linux_files/00-remix.sh -o /etc/profile.d/00-remix.sh
